pyproject.toml
[project]
name = "invest"
version = "0.1.0"
description = "Investment analysis and research tools"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "jupyter",
    "loguru",
    "yfinance>=0.2.50",
    "click",
    "python-dotenv",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "ib-async>=1.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/invest"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto" 
